+++ tests/installer-account-validation.sh
@@ -18,7 +18,10 @@ extract_function() {
 	case "$1" in
 		*[!abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_]*) return 1 ;;
 	esac
-	sed -n "/^$1() {\$/,/^}/p" "${SCRIPT_PATH}" >>"${FUNCTIONS_FILE}"
+	_extracted="${TMP_ROOT}/$1.function"
+	sed -n "/^$1() {\$/,/^}/p" "${SCRIPT_PATH}" >"${_extracted}" || return 1
+	[ -s "${_extracted}" ] || return 1
+	cat "${_extracted}" >>"${FUNCTIONS_FILE}"
 }
 
 extract_function adguardhome_owner_account
@@ -26,6 +29,7 @@ extract_function adguardhome_yaml_secure_file
 extract_function blocklist_yaml_replace_trap_disable
 extract_function blocklist_yaml_replace_trap_enable
 extract_function remove_unused_blocklists_from_yaml
+grep -q '^adguardhome_owner_account() {$' "${FUNCTIONS_FILE}" || fail 'adguardhome_owner_account extraction failed'
 sed 's#/bin/nvram#nvram#g; s#/usr/bin/awk#awk#g' "${FUNCTIONS_FILE}" >"${FUNCTIONS_FILE}.test" || fail 'could not isolate stock account commands'
 mv "${FUNCTIONS_FILE}.test" "${FUNCTIONS_FILE}" || fail 'could not update isolated account helpers'
 . "${FUNCTIONS_FILE}"
@@ -82,6 +86,7 @@ cmp -s "${YAML_FILE}.expected" "${YAML_FILE}" || fail 'ownership failure replace
 SERVICE_PATH="$(dirname "${SCRIPT_PATH}")/S99AdGuardHome"
 : >"${TMP_ROOT}/service-functions.sh"
 sed -n '/^adguardhome_owner_account() {$/,/^}/p' "${SERVICE_PATH}" >"${TMP_ROOT}/service-functions.sh"
+[ -s "${TMP_ROOT}/service-functions.sh" ] || fail 'service adguardhome_owner_account extraction failed'
 sed 's#/bin/nvram#nvram#g; s#/usr/bin/awk#awk#g' "${TMP_ROOT}/service-functions.sh" >"${TMP_ROOT}/service-functions.test" || fail 'could not isolate service stock account commands'
 mv "${TMP_ROOT}/service-functions.test" "${TMP_ROOT}/service-functions.sh" || fail 'could not update isolated service account helper'
 . "${TMP_ROOT}/service-functions.sh"
